Add all the fetch functions into one file
