+++ frontend/src/services/handlePlantApi.js
@@ -1,0 +1,71 @@
+const baseUrl = 'http://urbanplants.local/plant'
+
+export async function getPlants() {
+  try {
+    const response = await fetch(baseUrl)
+    const data = await response.json()
+    return await data
+  } catch (error) {
+    return { error: 'The server is down! :(' }
+  }
+}
+
+export async function postPlant(data) {
+  const myHeaders = new Headers()
+  myHeaders.append('Content-Type', 'application/json')
+  const copy = Object.assign({}, data)
+
+  const raw = JSON.stringify(copy)
+  const requestOptions = {
+    method: 'POST',
+    headers: myHeaders,
+    body: raw,
+    redirect: 'follow',
+  }
+  try {
+    const response = await fetch(baseUrl, requestOptions)
+    const responseData = response.json()
+    return responseData
+  } catch (error) {
+    return { error: 'The server is down! :(' }
+  }
+}
+
+export async function updatePlant(data, id) {
+  const myHeaders = new Headers()
+  myHeaders.append('Content-Type', 'application/json')
+  const copy = Object.assign({}, data)
+
+  const raw = JSON.stringify(copy)
+  const requestOptions = {
+    method: 'PUT',
+    headers: myHeaders,
+    body: raw,
+    redirect: 'follow',
+  }
+  try {
+    const response = await fetch(`${baseUrl}/${id}`, requestOptions)
+    const responseData = response.json()
+    return responseData
+  } catch (error) {
+    return { error: 'The server is down! :(' }
+  }
+}
+
+export async function deletePlant(id) {
+  const myHeaders = new Headers()
+  myHeaders.append('Content-Type', 'application/json')
+
+  const requestOptions = {
+    method: 'DELETE',
+    headers: myHeaders,
+    redirect: 'follow',
+  }
+  try {
+    const response = await fetch(`${baseUrl}/${id}`, requestOptions)
+    const responseData = response.json()
+    return responseData
+  } catch (error) {
+    return { error: 'The server is down! :(' }
+  }
+}
